Move ScaleCritiqueClient to a new module (#1634)
